requirements.txt
langchain-huggingface
langchain
langchain-core
langchain-pinecone
pdfservices-sdk
langchain-google-genai
pandas
langchain-community
python-dotenv
pdfminer.six
pinecone
os
dotenv